fix vulnerabilities
