+++ package.json
@@ -615,7 +615,7 @@
   },
   "scripts": {
     "vscode:prepublish": "webpack --mode production",
-    "build-plugin": "gulp build-plugin",
+    "build-plugin": "node scripts/build-jdtls-ext.js",
     "compile": "tsc -p ./",
     "tslint": "tslint -t verbose --project tslint.json",
     "watch": "webpack --mode development --watch --info-verbosity verbose",
@@ -635,21 +635,19 @@
     "@types/xml-zero-lexer": "^2.1.0",
     "@types/xml2js": "^0.4.8",
     "glob": "^7.1.6",
-    "gulp": "^4.0.2",
-    "gulp-cli": "^2.3.0",
     "mocha": "7.1.1",
     "ts-loader": "^4.4.2",
     "tslint": "^5.20.1",
     "tslint-microsoft-contrib": "^5.0.1",
     "typescript": "^4.2.3",
     "vscode-test": "^1.5.2",
-    "webpack": "^4.46.0",
+    "webpack": "^5.44.0",
     "webpack-cli": "^3.3.11"
   },
   "dependencies": {
     "axios": "^0.21.1",
     "expand-home-dir": "0.0.3",
-    "fast-glob": "^2.2.7",
+    "fast-glob": "^3.2.7",
     "fs-extra": "^4.0.3",
     "get-port": "^4.2.0",
     "lodash": "^4.17.21",
